+++ CHANGELOG.md
@@ -5,9 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
 ### Changed
 ### Fixed
+### Added
 
 ## [Unreleased]
 ### Changed
@@ -16,6 +16,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - removed storeManager from entity resource constructor
 - changed context model in entity constructor
 - added getStore method to entity and get default store id from store manager
+### Added
+- attributes group codes
 
 ## [100.2.1] - 2019-08-29
 ### Fixed